# Look Ma, no `@generated!`