--- a/Project.toml
+++ b/Project.toml
@@ -17,7 +17,7 @@ Polyester = "0.5, 0.6, 0.7"
 Static = "0.3, 0.4, 0.5, 0.6, 0.7, 0.8"
 StaticArrayInterface = "1"
 StrideArraysCore = "0.3, 0.4, 0.5"
-julia = "1.6"
+julia = "1.8"
 
 [extras]
 InteractiveUtils = "b77e0a4c-d291-57a0-90e8-8db25a27a240"