First thing fist, master:
```julia
julia> @time using OrdinaryDiffEq
Precompiling OrdinaryDiffEq
  1 dependency successfully precompiled in 46 seconds. 155 already precompiled.
 51.330664 seconds (8.43 M allocations: 598.668 MiB, 0.47% gc time, 6.02% compilation time)

(@fbtestrelease) pkg> st -m OrdinaryDiffEq FastBroadcast DiffEqBase
Status `~/.julia/environments/fbtestrelease/Manifest.toml`
  [2b5f629d] DiffEqBase v6.151.1
  [7034ab61] FastBroadcast v0.2.8
  [1dea7af3] OrdinaryDiffEq v6.78.0 `~/.julia/dev/OrdinaryDiffEq`
```
PR
```julia
julia> @time using OrdinaryDiffEq
Precompiling OrdinaryDiffEq
  1 dependency successfully precompiled in 43 seconds. 155 already precompiled.
 48.351458 seconds (8.34 M allocations: 592.818 MiB, 0.46% gc time, 6.02% compilation time)

(@fbtest) pkg> st -m OrdinaryDiffEq FastBroadcast DiffEqBase
Status `~/.julia/environments/fbtest/Manifest.toml`
  [2b5f629d] DiffEqBase v6.151.1
  [7034ab61] FastBroadcast v0.2.8 `~/.julia/dev/FastBroadcast`
  [1dea7af3] OrdinaryDiffEq v6.78.0 `~/.julia/dev/OrdinaryDiffEq`
```
so, this doesn't really help much. But the time improvement may be real, e.g. I get this fairly consistently when running tests (and setting the CPU governor to perf)
Master:
```julia
julia> @time using FastBroadcast, PerformanceTestTools, MuladdMacro
  0.110391 seconds (108.60 k allocations: 7.431 MiB, 18.06% compilation time)

julia> FastBroadcast.DEBUG[]=true # I'm running the new test file, which removed this line
true

julia> @time include("test/runtests.jl")
Test Summary: | Pass  Total  Time
FastBroadcast |   42     42  6.9s
  8.446610 seconds (11.79 M allocations: 781.170 MiB, 2.99% gc time, 99.19% compilation time: 1% of which was recompilation)
```
vs PR
```julia
julia> @time using FastBroadcast, PerformanceTestTools, MuladdMacro
  0.094999 seconds (91.83 k allocations: 6.864 MiB, 13.47% compilation time)

julia> @time include("test/runtests.jl")
Test Summary: | Pass  Total  Time
FastBroadcast |   42     42  6.2s
  7.765427 seconds (16.93 M allocations: 1.093 GiB, 4.33% gc time, 98.95% compilation time: 1% of which was recompilation)
```
Note that the allocations increased a lot, but the time improved.

Maybe I should write a loop that `touch`es `OrdinaryDiffEq.jl`, and repeatedly re-precompiles to get an idea of how consistent this 3s saving is?

Anyway, about the change of approach here:
The original motivation for the `@generated` approach was that we didn't trust the compiler to peel through all the abstractions.

Base actually no longer allocates for these test examples:
```julia
using FastBroadcast, MuladdMacro
function foo9_fb(a, b, c, d, e, f, g, h, i)
    @muladd @.. a = b + 0.1 * (0.2c + 0.3d + 0.4e + 0.5f + 0.6g + 0.6h + 0.6i)
    nothing
end
function foo9_sb(a, b, c, d, e, f, g, h, i)
    @muladd @. a = b + 0.1 * (0.2c + 0.3d + 0.4e + 0.5f + 0.6g + 0.6h + 0.6i)
    nothing
end

function foo26_fb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
    @muladd @.. a =
        b +
        0.1 * (
            0.2c +
            0.3d +
            0.4e +
            0.5f +
            0.6g +
            0.6h +
            0.6i +
            0.6j +
            0.6k +
            0.6l +
            0.6m +
            0.6n +
            0.6o +
            0.6p +
            0.6q +
            0.6r +
            0.6s +
            0.6t +
            0.6u +
            0.6v +
            0.6w +
            0.6x +
            0.6y +
            0.6z
        )
    nothing
end
function foo26_sb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
    @muladd @. a =
        b +
        0.1 * (
            0.2c +
            0.3d +
            0.4e +
            0.5f +
            0.6g +
            0.6h +
            0.6i +
            0.6j +
            0.6k +
            0.6l +
            0.6m +
            0.6n +
            0.6o +
            0.6p +
            0.6q +
            0.6r +
            0.6s +
            0.6t +
            0.6u +
            0.6v +
            0.6w +
            0.6x +
            0.6y +
            0.6z
        )
    nothing
end

a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z =
    [rand(100, 100, 2) for i = 1:26];
@time @eval foo9_sb(a, b, c, d, e, f, g, h, i)
@time @eval foo9_fb(a, b, c, d, e, f, g, h, i)
@time @eval foo26_sb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
@time @eval foo26_fb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)

@btime foo9_sb($a, $b, $c, $d, $e, $f, $g, $h, $i)
@btime foo9_fb($a, $b, $c, $d, $e, $f, $g, $h, $i)
@btime foo26_sb($a, $b, $c, $d, $e, $f, $g, $h, $i, $j, $k, $l, $m, $n, $o, $p, $q, $r, $s, $t, $u, $v, $w, $x, $y, $z)
@btime foo26_fb($a, $b, $c, $d, $e, $f, $g, $h, $i, $j, $k, $l, $m, $n, $o, $p, $q, $r, $s, $t, $u, $v, $w, $x, $y, $z)
```
so they've sorted things out there; PR:
```julia
julia> @btime foo9_sb($a, $b, $c, $d, $e, $f, $g, $h, $i)
  81.820 μs (0 allocations: 0 bytes)

julia> @btime foo9_fb($a, $b, $c, $d, $e, $f, $g, $h, $i)
  35.181 μs (0 allocations: 0 bytes)

julia> @btime foo26_sb($a, $b, $c, $d, $e, $f, $g, $h, $i, $j, $k, $l, $m, $n, $o, $p, $q, $r, $s, $t, $u, $v, $w, $x, $y, $z)
  462.470 μs (0 allocations: 0 bytes)

julia> @btime foo26_fb($a, $b, $c, $d, $e, $f, $g, $h, $i, $j, $k, $l, $m, $n, $o, $p, $q, $r, $s, $t, $u, $v, $w, $x, $y, $z)
  136.171 μs (0 allocations: 0 bytes)
```
It's basically the same on master.


Compile times did change here; master we had:
```julia
julia> @time @eval foo9_sb(a, b, c, d, e, f, g, h, i)
  0.494337 seconds (1.19 M allocations: 79.877 MiB, 4.67% gc time, 99.93% compilation time)

julia> @time @eval foo9_fb(a, b, c, d, e, f, g, h, i)
  0.318403 seconds (357.25 k allocations: 23.489 MiB, 99.88% compilation time)

julia> @time @eval foo26_sb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
  1.084712 seconds (6.43 M allocations: 404.590 MiB, 14.40% gc time, 99.92% compilation time)

julia> @time @eval foo26_fb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
  0.248650 seconds (691.49 k allocations: 42.912 MiB, 99.73% compilation time)
```
vs PR:
```julia
julia> @time @eval foo9_sb(a, b, c, d, e, f, g, h, i)
  0.474962 seconds (1.19 M allocations: 79.877 MiB, 4.78% gc time, 99.92% compilation time)

julia> @time @eval foo9_fb(a, b, c, d, e, f, g, h, i)
  0.131016 seconds (641.46 k allocations: 43.612 MiB, 8.18% gc time, 99.73% compilation time)

julia> @time @eval foo26_sb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
  1.024314 seconds (6.43 M allocations: 404.590 MiB, 14.07% gc time, 99.92% compilation time)

julia> @time @eval foo26_fb(a, b, c, d, e, f, g, h, i, j, k, l, m, n, o, p, q, r, s, t, u, v, w, x, y, z)
  0.532058 seconds (4.14 M allocations: 265.577 MiB, 8.14% gc time, 99.87% compilation time)
```
That the second compilation is faster on master than in this PR is true even if we switch orders. That is a little concerning.

It's up in the air whether this approach is an improvement, but if the 3s improvement for OrdinaryDiffEq is real, that's not much but at least something.

`@generated` is a hack, but so is the `recursion_relation = Returns(true)` and large number of base methods I needed to re-implement here to get them type stable, as the compiler continues to handle these things poorly.
Readers can decide which approach is more readable.

We could drop some of the `@inline`s, but there is a test that `@code_llvm` shows a vector body. That requires inlining at least some methods.
Also inlining the top level `fast_materialize(!)` is unnecessary for that, but it saves us from stack allocating the object, and (because we're inlining with `@propagate_inbounds`) means we can remove the error check, to be more GPUCompiler-friendly.

Not checking bounds at all was the default (`@inbounds` or not) on master. I could restore that behavior.

@ChrisRackauckas